backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-multipart>=0.0.9
pillow>=10.0.0
numpy>=1.24.0
tensorflow>=2.13.0
huggingface_hub>=0.22.0
python-dotenv>=1.0.0
